Successfully debugged toggleNav and isNavVisible functionality.
