+++ src/styles/App.css
@@ -13,25 +13,77 @@
     border-top: none;
   }
 
+  .navLinks {
+    position: fixed;
+    right: 0;
+    top: 0;
+    left: 0;
+    background-color: black;
+    width: 100%;
+    display: flex;
+    height: 7.5%;
+    justify-content: center;
+    text-align: center;
+    align-items: center;
+    border-bottom: none;
+    z-index: 1000;
+    padding: 0 20px;
+  }
+
+  .navLinks.hidden {
+    display: none;
+  }
+
+.linkAnchor:visible {
+  color: white;
+  text-decoration: none;
+
+}
+
+  .navLinks>a,
+  .navLinks .linkGroup>a {
+    font-size: 20px;
+    padding: 10px;
+    margin-left: 5%;
+    color: white;
+    text-decoration: underline;
+    font-weight: bold;
+  }
+
+  .navLinks>a:hover,
+  .navLinks .linkGroup>a:hover {
+    color: #39ff14;
+    transform: scale(1.1);
+  }
+
+  .closeNavButton {
+    background: none;
+    position: fixed;
+    border: solid white 3px;
+    border-radius: 25%;
+    font-size: 35px;
+    cursor: pointer;
+    color: white;
+    outline: none;
+    margin: 1%;
+    padding-top: 0px;
+    top: 0;
+    left: 0;
+    z-index: 0;
+  }
+
+  .closeNavButton:hover {
+    background-color: black;
+    color: #39ff14;
+    border: solid white 2px;
+  }
+
   form {
     display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: center;
     width: 100%;
-  }
-
-  .cookieConsent {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: black;
-    color: white;
-    font-size: 15px;
-    padding: 10px;
-    border-top: solid grey 3px;
-    border-bottom: solid grey 3px;
-    position: center;
   }
 
   html {
